backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1

# RAG & AI (pure Python, no C++/Rust/PyTorch required)
langchain-core>=0.3.0
langchain-groq>=0.2.0
scikit-learn>=1.3.0

# Development
tiktoken>=0.7.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
